Initial Weighment Float

+++ src/main/java/com/sericulture/model/LotPendingReportResponse.java
@@ -39,7 +39,7 @@ public class LotPendingReportResponse {
     private float lotSoldOutAmount;
     private float bidAmount;
     private double reelerCurrentBalance;
-    private int farmerEstimatedWeight;
+    private Float farmerEstimatedWeight;
     private List<Float> lotWeightDetail;
     private String marketName;
     private String race;
